fix: from pr comments
