+++ libs/shared/src/entities/feed/feed.interface.ts
@@ -1,0 +1,6 @@
+export interface IFeedEntity {
+  _id: string;
+  name: string;
+  _environmentId: string;
+  _organizationId: string;
+}
