+++ libs/shared/src/index.ts
@@ -4,6 +4,7 @@ export * from './entities/organization';
 export * from './entities/notification-template';
 export * from './entities/environment';
 export * from './entities/messages/messages.interface';
+export * from './entities/feed/feed.interface';
 export * from './types/builder/builder.types';
 export * from './entities/notification';
 export * from './entities/message-template';
