+++ packages/notification-center/src/components/notification-center/components/FeedsTabs.tsx
@@ -29,7 +29,7 @@ export function FeedsTabs() {
           ))}
         </Tabs>
       ) : (
-        <NotificationsListTab feedId={''} />
+        <NotificationsListTab />
       )}
     </>
   );
